Refactor movie API server with improved filtering and recommendations
